# Optimize compress_block() and build_tree()