@mtl1979 2025-08-19
The failures on macOS workflows seem to be caused by mismatch between gcc/clang version and Xcode version. Xcode 16 is currently oldest supported but some workflows try to use for example Xcode 15.2.